refactor: 2단계 요구사항 반영
domain 로직과 view 로직을 분리하였음.
