+++ src/Controller/Controller.js
@@ -1,7 +1,7 @@
 const InputView = require("../View/InputView");
 const OutputView = require("../View/OutputView");
-const ValidateCarName = require("../ValidateCarName");
-const ValidateCountOfTrial = require("../ValidateCountOfTrial");
+const ValidateCarName = require("../utils/ValidateCarName");
+const ValidateCountOfTrial = require("../utils/ValidateCountOfTrial");
 const Race = require("../Domain/Race");
 
 class Controller {
